crearte order model and controller

+++ src/models/cart.model.ts
@@ -1,15 +1,8 @@
-import express, { Router, Request, Response } from "express";
 import { PrismaClient } from "@prisma/client";
-import bcrypt from "bcrypt"
 import dotenv from "dotenv"
-import { exitCode } from "process";
 dotenv.config();
 
-const app = express();
 const prisma = new PrismaClient();
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
 class CartModel {
 
     async createCart(userId: string) {
@@ -56,7 +49,7 @@ class CartModel {
         try {
             const cart = await prisma.cart.findUnique({
                 where: {
-                    id: id
+                    id
                 }
             });
             return cart;
@@ -65,23 +58,27 @@ class CartModel {
             throw new Error(`Error retrieving the cart ${(error as Error).message}`);
         }
     }
-    async deleteCartById(req: Request) {
+    async deleteCartById(id:string) {
         try {
-            const id: string = req.params.id;
-
-            const cart = await prisma.cart.delete({
-                where: {
-                    id: id
+            const cartProduct = await prisma.cartProduct.deleteMany({
+                where:{
+                    cartId:id
                 }
-            });
-            if (!cart)
-                throw new Error(`no user found with that id`);
-            return cart;
+            })
+            if(cartProduct){
+                const cart = await prisma.cart.delete({
+                    where: {
+                        id
+                    }
+                });
+                if (!cart)
+                    throw new Error(`no cart found with that id`);
+                return cart;
+            }
         } catch (error) {
-            throw new Error(`Error deleting the user ${(error as Error).message}`);
+            throw new Error(`Error deleting the cart ${(error as Error).message}`);
         }
     }
 }
-
 
 export default CartModel;
